+++ workflow/bin/start.sh
@@ -1,8 +1,15 @@
-#!/bin/sh
+#!/bin/bash
 
 #
 # Startup script for Workflow server
 #
+
+if [ "$EUID" -eq 0 ]
+then
+    echo "**** Please do NOT run UNICORE Workflow as 'root'!"
+    echo "     It's a bad security practice."
+    exit 1
+fi
 
 #
 # Installation directory